Increase Creative Cell capacity from int to long (#580)

+++ src/main/java/appeng/me/storage/CreativeCellInventory.java
@@ -34,7 +34,7 @@ public class CreativeCellInventory implements IMEInventoryHandler<IAEItemStack> 
         for (final ItemStack is : cc) {
             if (is != null) {
                 final IAEItemStack i = AEItemStack.create(is);
-                i.setStackSize(Integer.MAX_VALUE);
+                i.setStackSize(Long.MAX_VALUE);
                 this.itemListCache.add(i);
             }
         }
